estilizando edit form

+++ clothes-store/src/app/all-tools/products-tool/products-tool.component.ts
@@ -5,16 +5,17 @@ import { AsyncPipe, NgIf } from '@angular/common';
 import { Router, RouterLink } from '@angular/router';
 import { CreatingProductComponent } from './creating-product/creating-product.component';
 import { Observable } from 'rxjs';
+import { FormsModule } from '@angular/forms';
 
 @Component({
   selector: 'app-products-tool',
-  imports: [NgIf, CreatingProductComponent, AsyncPipe],
+  imports: [NgIf, CreatingProductComponent, AsyncPipe, FormsModule],
   templateUrl: './products-tool.component.html',
   styleUrl: './products-tool.component.css'
 })
 export default class ProductsToolComponent implements OnInit{
 
-  // allProducts: productModule[] = [] 
+
   allProducts$ = new Observable<productModule[]>()
 
   productService = inject(ProductsService)
@@ -27,9 +28,7 @@ export default class ProductsToolComponent implements OnInit{
   }
 
   fetchingAllProducts(){
-    // this.productService.allProducts$.subscribe((res) =>{
-    //   this.allProducts = res
-    // })
+    
     this.allProducts$ = this.productService.allProducts$
   }
 
@@ -47,16 +46,32 @@ export default class ProductsToolComponent implements OnInit{
   }
 
 
+
+  // editing
+  editDialogOpen = true
+
+  editProduct(item: any){
+
+  }
+
+
   // deleting product
   deleteProduct(id: number){
-    this.productService.deleteProduct(id).subscribe({ // window.location.reload()
-      next: () => {
-        console.log('Product deleted')
-        
-      },
-      error: (err) => console.log('error', err)
-    })
+
+    if(window.confirm('Are you sure you wanna delete this item?')){
+
+      this.productService.deleteProduct(id).subscribe({
+        next: () => {
+          console.log('Product deleted')
+          window.location.reload()
+        },
+        error: (err) => console.log('error', err)
+      })
+    }
+
   }
+
+
